requirements.txt

pandas==1.3.4
python-dotenv==0.19.2
email-to==0.1.0
numpy==1.21.6
requests==2.26.0
DateTime==5.0
pymssql==2.2.5
azure-storage-blob==2.1.0